pyproject.toml
[project]
name = "galaxy-mail-analyzer"
version = "0.1.0"
description = "Analyze galaxy-bugs mailing list emails with AI-powered classification and knowledge base matching"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "Galaxy Project" }
]
keywords = ["galaxy", "bioinformatics", "email", "analysis", "AI"]

dependencies = [
    # Database
    "sqlalchemy>=2.0",
    "alembic>=1.13",

    # Vector store
    "chromadb>=0.4",

    # AI/ML
    "voyageai>=0.3",
    "anthropic>=0.40",
    "bertopic>=0.16",
    "hdbscan>=0.8.33",
    "umap-learn>=0.5",

    # HTTP client
    "httpx>=0.27",

    # Configuration and CLI
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "typer>=0.12",
    "rich>=13.0",
    "python-dotenv>=1.0",

    # Text processing
    "beautifulsoup4>=4.12",
    "html2text>=2024.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.1",
    "mypy>=1.8",
    "ruff>=0.3",
]
postgres = [
    "psycopg[binary]>=3.1",
]

[project.scripts]
galaxy-mail = "galaxy_mail_analyzer.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/galaxy_mail_analyzer"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
